@@ -1,11 +1,11 @@
 package kyamls
 
 import (
+	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
 
-	"github.com/pkg/errors"
 	"sigs.k8s.io/kustomize/kyaml/yaml"
 )
 
@@ -13,7 +13,7 @@ import (
 func ModifyFiles(dir string, modifyFn func(node *yaml.RNode, path string) (bool, error), filter Filter) error {
 	filterFn, err := filter.ToFilterFn()
 	if err != nil {
-		return errors.Wrap(err, "failed to create filter")
+		return fmt.Errorf("failed to create filter: %w", err)
 	}
 
 	err = filepath.Walk(dir, func(path string, info os.FileInfo, err error) error {
@@ -25,13 +25,13 @@ func ModifyFiles(dir string, modifyFn func(node *yaml.RNode, path string) (bool,
 		}
 		node, err := yaml.ReadFile(path)
 		if err != nil {
-			return errors.Wrapf(err, "failed to load file %s", path)
+			return fmt.Errorf("failed to load file %s: %w", path, err)
 		}
 
 		if filterFn != nil {
 			flag, err := filterFn(node, path)
 			if err != nil {
-				return errors.Wrapf(err, "failed to evaluate filter on file %s", path)
+				return fmt.Errorf("failed to evaluate filter on file %s: %w", path, err)
 			}
 			if !flag {
 				return nil
@@ -40,7 +40,7 @@ func ModifyFiles(dir string, modifyFn func(node *yaml.RNode, path string) (bool,
 
 		modified, err := modifyFn(node, path)
 		if err != nil {
-			return errors.Wrapf(err, "failed to modify file %s", path)
+			return fmt.Errorf("failed to modify file %s: %w", path, err)
 		}
 
 		if !modified {
@@ -49,12 +49,12 @@ func ModifyFiles(dir string, modifyFn func(node *yaml.RNode, path string) (bool,
 
 		err = yaml.WriteFile(node, path)
 		if err != nil {
-			return errors.Wrapf(err, "failed to save %s", path)
+			return fmt.Errorf("failed to save %s: %w", path, err)
 		}
 		return nil
 	})
 	if err != nil {
-		return errors.Wrapf(err, "failed to modify files in dir %s", dir)
+		return fmt.Errorf("failed to modify files in dir %s: %w", dir, err)
 	}
 	return nil
 }
