@@ -1,7 +1,8 @@
 package kyamls
 
 import (
-	"github.com/pkg/errors"
+	"fmt"
+
 	"sigs.k8s.io/kustomize/kyaml/yaml"
 )
 
@@ -9,7 +10,7 @@ import (
 func SetStringValue(node *yaml.RNode, path string, value string, fields ...string) error {
 	err := node.PipeE(yaml.LookupCreate(yaml.ScalarNode, fields...), yaml.FieldSetter{StringValue: value})
 	if err != nil {
-		return errors.Wrapf(err, "failed to set field %s to %s at path %s", JSONPath(fields...), value, path)
+		return fmt.Errorf("failed to set field %s to %s at path %s: %w", JSONPath(fields...), value, path, err)
 	}
 	return nil
 }
