@@ -2,10 +2,9 @@ package stringhelpers
 
 import (
 	"bytes"
+	"fmt"
 	"net/url"
 	"strings"
-
-	"github.com/pkg/errors"
 )
 
 // UrlJoin joins the given paths so that there is only ever one '/' character between the paths
@@ -92,7 +91,7 @@ func SanitizeURL(unsanitizedUrl string) string {
 func URLSetUserPassword(rawURL, username, password string) (string, error) {
 	u, err := url.Parse(rawURL)
 	if err != nil {
-		return rawURL, errors.Wrapf(err, "failed to parse URL %s", rawURL)
+		return rawURL, fmt.Errorf("failed to parse URL %s: %w", rawURL, err)
 	}
 	user := u.User
 	if user != nil {
