@@ -1,6 +1,8 @@
 package helper
 
-import "fmt"
+import (
+	"fmt"
+)
 
 // BashExample returns markdown for a bash script expression for the CLI help
 func BashExample(binaryCommand string, cli string) string {
