@@ -1,6 +1,8 @@
 package scmhelpers
 
-import "github.com/jenkins-x/go-scm/scm"
+import (
+	"github.com/jenkins-x/go-scm/scm"
+)
 
 // ContainsLabel returns true if the given labels contains the given label name
 func ContainsLabel(labels []*scm.Label, label string) bool {
